--- Moriquendi/ClassMF/ClassMF_ocr/__init__.py
@@ -1,1 +1,0 @@
-#exemplo para criar um pacote novo de funções aqui
