+++ Moriquendi/mfglobal.py
@@ -7,13 +7,9 @@
     """
 
 # Inicio --> Variáveis de Cadastro de dados
-
-from itertools import count
-
-
 NomePlayer = ["andreribei00","andreribeir0","AquelaSereia","BangTan 7","Ben Reilly","BlackWidowMF","BR WOLVERINE","Catotinha","dhyessika","Duffly","Fabiiano","FoxPop","Green Vile","hvReus","JEFFER132","Junior crvg","KRATOS 79 BR","KRATOS BR MF","Lady Catota","lalinha44","Lightside 1","Lord Costela","LordRafaeel","MadMax420","MadMoustache","MeniPlays","isaveio","Mx Julielton","Nioljso","O Pae Ta On","princepesam","Rainha Agnes","Rasanzi,Sammy A","Smith Master","sunsupershin","UM DOIS 1 2","Vanda mob","willrick001","Wolverine BR","xX pipoca Xx","xX SAC Xx","xXsaquinXx"] #Lista de membros da guild, separados.. cada nome um item da lista
 #NomePlayer é uma lista dos jogadores cadastrados
-qt = len(NomePlayer)   #Soma a quantidade de nomes cadastrados em NomePlayer
+qt =+ len(NomePlayer)   #Soma a quantidade de nomes cadastrados em NomePlayer
 
 IdJogador = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 
 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 
@@ -25,9 +21,6 @@ IdJogador = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 
 # Inicio --> Variáveis do Menu Inicial
 
 contador = 0    #contador soma quantas opções do menu é escolhida
-digop = ""      #armazena nome do jogador para pesquisa na lista
-
-opcao = "0"     #armazena a opção escolhida no menu principal, temporariamente. String para evitar erros quando digitado caracteres ou algo diferente de numeros
 
 # FIM --> Variáveis do Menu Inicial
