+++ Moriquendi/testeMF.py
@@ -1,8 +1,9 @@
 #-*- coding: -utf8 -*-
-from itertools import count
-from VarGlobal import IdJogador
-from audioop import add
+from classMF.packMF import ListaPlayer, MenuCad
 import os
 import sys
-import ClassMF #Importa pacote de funções criados em uma pasta separada
+#from ClassMF.__PackFun__ import * #Importa pacote de funções criados em uma pasta separada
 
+from mfglobal import *
+
+MenuCad()
